requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.29,<1
sqlmodel>=0.0.22,<0.0.99
SQLAlchemy>=2.0.25,<3
httpx>=0.27,<1
jinja2>=3.1,<4
fastapi
httpx
SQLAlchemy
sqlmodel
uvicorn[standard]
python-dotenv
ccxt
numpy
pandas
python-dateutil
xgboost
